+++ src/ModularPipelines/DependencyInjection/DependencyInjectionSetup.cs
@@ -67,6 +67,8 @@ internal static class DependencyInjectionSetup
         // Singletons
         services
             .AddSingleton<IConsolePrinter, ConsolePrinter>()
+            .AddSingleton<AfterPipelineLogger>()
+            .AddSingleton<IPipelineGlobalHooks>(x => x.GetRequiredService<AfterPipelineLogger>())
             .AddSingleton<IExceptionContainer, ExceptionContainer>()
             .AddSingleton<IPipelineContextProvider, ModuleContextProvider>()
             .AddSingleton<IDependencyChainProvider, DependencyChainProvider>()