+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
 *   Throw an exception when defining duplicate constraint keys on the same NotInParallel attribute
+*   `context.LogOnPipelineEnd(...)` method to easily log information at the end of the pipeline where it's easier to find in logs